Sync origin master with upstream master (#1)
* Create dockerfile to run pshtt in a container

  * Basic dockerfile extending python:2.7-onbuild
  * Create custom entrypoint to change ownership of results file
  * Add a bit of documentation to the README
    - how to build the image
    - how to use it

* Update README.md

Moving Docker explainer text lower

* Distinguish between connection errors and other exceptions.

* Refactor into a package

Note that Python has an unfortunate naming collision here. This commit
refactors the existing pshtt code into a "package" - that is, a
multi-file module with an `__init__.py`. This is distinct from a
"package on PyPi" (which is a "distribution package"). However, this
refactoring does make it easier to write a `setup.py` which can be used
to create a package for PyPi.

This refactor changes the calling convention for the pshtt CLI tool from
`./pshtt_cli` to `python -m pshtt.cli`. The Docker entrypoint script and
the README are updated accordingly.

* setup.py for installing pshtt

You can run `python setup.py install` to install pshtt (and its
dependencies). You should also be able to use `setup.py` to create a
distribution package suitable for uploading to PyPi.

Note that `setup.py` defines `entry_points`, which makes the pshtt CLI
tool available simply as `pshtt`. Nice!

* fix typo

* Loosen dependency restrictions

* Recommend pip development mode install

* Fix lint errors

* Update setup.py

Author update, license parity